Disable keyboard shortcuts in new private conversation message box
It is not uncommon for the message box to pop up while the user is in the
middle of writing a conversation. Thus, it is best to disable the default Y/N
shortcuts which might cause the request to be wrongly accepted or rejected due
to keys being incorrectly sent to the message box.