Implement proper handling for server-sent code 1003 messages
The server seems to send a code 1003 message every 4 minutes to the client. Upon
reception, the client resets its timer and sends a 1003 response immediately.
Previous versions ignored this message completely and it didn't seem to cause
any harm, but let's do what the official client does from now on.