Fix memory leaks created by mishandling login session objects
Unfortunately, a LoginSession object cannot be a QSharedPointer right from the
start. This is because, if the session fails, the object will never be
destroyed, because the lambda which is assigned to loginSuccessful will keep
holding onto its own instance of QSharedPointer until the object which emits the
signal (which is incidentally *the same* LoginSession object) is destroyed.
Creating the LoginSession object with ordinary new and wrapping it with a
QSharedPointer only if the login was successful - and the session can thus be
managed and used by chat window - solves the issue, but obviously means that
we're unfortunately back to the world of manual lifetime management.

Obviously, this signals a bigger issue : there should be a separate facility
which hands out LoginSession objects only if they are actually usable, so that
the client code doesn't need to bother with deleting a "failed" LoginSession
that isn't usable later anyway. A stateless global "session initiator" QObject
with login{Successful,Failed} and captchaRequired signals, where loginSuccessful
provides an exclusive pointer to a "live" LoginSession would probably do the
trick.

This could serve as an "actual code" reason to avoid zombie objects where
possible.