(Hopefully) fix crash on reconnecting when using sessions created by autologin
Well, this is embarassing.

Even though AutologinState is only created once at creation of MainWindow, the
lambdas connected to the signals of the created sessions outlive the
AutologinState object, which leads to a crash when reconnecting due to
loginSuccessful being fired long after the AutologinState object is destroyed.
The crash occurs because the lambda captures a AutologinState *this in order to
access the mMainWindow member.