Don't call QMessageBox::reject() in NotifSupMsgBox::removeNotification
Calling reject() indirectly calls onPrivateConvNotificationRejected() via the
finished() signal connected in displayNotification(). It may have seemed like a
good idea at the time of writing this code, but it seems like this might be
somehow related to the mysterious "conversation tab gets closed out of nowhere"
bug that's triggered in a rather specific case described in #81.

None of the other notification implementations actually do anything which would
interact with the network or the rest of the UI, simply limiting themselves to
removing the notification from the list of "alive" ones. Inspecting
pre-NotifSupport code also shows that the implementation of
QMessageBox::finished did not do any of the stuff that
onPrivateConvNotificationRejected() does.