Minor image processing improvements
The image data received from the server will now be saved directly to the disk
if the user so chooses : this avoids an unnecessary decode/encode cycle enforced
by the creation of a QImage, potentially reducing the quality of the image.
A QImage object is now created only for the purpose of showing the image.